Module pallet_xcm_bridge_router::pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

The main components of this pallet are:

Pallet, which implements all of the dispatchable extrinsics of the pallet, among other public functions.

The subset of the functions that are dispatchable can be identified either in the dispatchables module or in the Call enum.

storage_types, which contains the list of all types that are representing a storage item. Otherwise, all storage items are listed among Type Definitions.

Config, which contains the configuration trait of this pallet.

Event and [Error], which are listed among the Enums.

Re-exports

pub use __tt_error_token_1 as tt_error_token;

pub use __tt_default_parts_7 as tt_default_parts;

pub use __tt_extra_parts_7 as tt_extra_parts;

pub use __tt_default_parts_v2_7 as tt_default_parts_v2;

Modules

config_preludes

Default implementations of DefaultConfig, which can be used to implement Config.

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Structs

Instance1

Instance1 to be used for instantiable pallets defined with the #[pallet] macro. Instances 2-16 are also available but are hidden from docs.

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Event

The Event enum of this pallet

Traits

Config

Configuration trait of this pallet.

DefaultConfig

Based on Config. Auto-generated by #[pallet::config(with_default)]. Can be used in tandem with #[register_default_config] and #[derive_impl] to derive test config traits based on existing pallet config traits in a safe and developer-friendly way.

Type Aliases

BridgeIdOf

An alias for the BridgeId of configured T::BridgeIdResolver.

Bridges

Stores BridgeState for congestion control and dynamic fees for each resolved bridge ID associated with a destination.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.